+++ bcbio/pipeline/datadict.py
@@ -86,6 +86,8 @@ LOOKUPS = {
     "offtarget_stats": {"keys": ["regions", "offtarget_stats"]},
     "sample_callable": {"keys": ["regions", "sample_callable"]},
     "coverage_interval": {"keys": ["config", "algorithm", "coverage_interval"]},
+    "coverage_experimental": {"keys": ["config", "algorithm", "coverage_experimental"]},
+    "report": {"keys": ["config", "algorithm", "report"]},
     "coverage_depth_min": {"keys": ["config", "algorithm", "coverage_depth_min"],
                            "default": 4},
     "coverage_depth_max": {"keys": ["config", "algorithm", "coverage_depth_max"],