+++ bcbio/pipeline/main.py
@@ -190,6 +190,11 @@ class Variant2Pipeline(AbstractPipeline):
                 samples = population.prep_db_parallel(samples, run_parallel)
             with profile.report("quality control", dirs):
                 samples = qcsummary.generate_parallel(samples, run_parallel)
+
+        with prun.start(_wres(parallel, ["gatk"], ensure_mem = {"gatk": 8}),
+                        samples, config, dirs, "coverage") as run_parallel:
+            with profile.report("report", dirs):
+                samples = qcsummary.report_summary(samples, run_parallel)
             with profile.report("archive", dirs):
                 samples = archive.compress(samples, run_parallel)
             with profile.report("upload", dirs):
@@ -240,10 +245,12 @@ class StandardPipeline(AbstractPipeline):
                 samples = run_parallel("combine_sample_regions", [samples])
                 samples = region.clean_sample_data(samples)
         ## Quality control
-        with prun.start(_wres(parallel, ["fastqc", "bamtools", "samtools", "qsignature", "kraken"]),
+        with prun.start(_wres(parallel, ["fastqc", "bamtools", "qsignature", "kraken", "gatk"], ensure_mem={"gatk" : 2}),
                         samples, config, dirs, "multicore2") as run_parallel:
             with profile.report("quality control", dirs):
                 samples = qcsummary.generate_parallel(samples, run_parallel)
+            with profile.report("report", dirs):
+                samples = qcsummary.report_summary(samples, run_parallel)
             with profile.report("upload", dirs):
                 samples = run_parallel("upload_samples", samples)
                 for sample in samples: