+++ HISTORY.md
@@ -7,6 +7,8 @@
   Schroeder.
 - Initial support for small RNA pipeline. Thanks to Lorena Pantano.
 - Support for MetaSV to prepare combined structural variant calls.
+- Add smallRNA-seq pipeline
+- Test automatic report for variants calling and standard pipeline.
 
 ## 0.9.1 (6 August 2015)